int64

+++ health.go
@@ -10,7 +10,7 @@ package main
     //import "github.com/ssimunic/gosensors"
     //import "github.com/kolide/osquery-go"
 
-func read_sensors() int{
+func read_sensors() int64{
 	//out, err := exec.Command("cat /sys/bus/platform/devices/coretemp.0/hwmon/hwmon2/temp2_input").Output()
 	filerc, err := os.Open("/sys/bus/platform/devices/coretemp.0/hwmon/hwmon2/temp2_input")
 	if err != nil {
